+++ .github/workflows/test.yml
@@ -27,9 +27,5 @@ jobs:
         uses: ./
         with:
           working-directory: "./tests/mockReports"
+          github-token: ${{ secrets.GITHUB_TOKEN }}
           name: "Mock Reports"
-        env:
-          # See https://github.com/changesets/action/issues/147
-          HOME: ${{ github.workspace }}
-          # Allows to download / query / comment packages
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
