+++ src/inputs/getGithubToken.ts
@@ -1,0 +1,6 @@
+import * as core from '@actions/core';
+
+export const getGithubToken = () => {
+  const gitHubToken = core.getInput('github-token').trim();
+  return gitHubToken;
+};
